requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via
    #   langchain-community
    #   litellm
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   langfuse
    #   openai
argon2-cffi==23.1.0
    # via minio
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via langfuse
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   minio
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   litellm
    #   streamlit
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
greenlet==3.2.0
    # via sqlalchemy
grpcio==1.67.1
    # via pymilvus
h11==0.14.0
    # via httpcore
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via
    #   langfuse
    #   langsmith
    #   litellm
    #   openai
httpx-sse==0.4.0
    # via langchain-community
huggingface-hub==0.30.2
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   langfuse
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via litellm
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   litellm
    #   pydeck
jiter==0.9.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via
    #   altair
    #   litellm
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.23
    # via langchain-community
langchain-community==0.3.21
    # via -r requirements.in
langchain-core==0.3.54
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.14
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langfuse==2.60.3
    # via -r requirements.in
langsmith==0.3.32
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
litellm==1.67.0
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
minio==7.2.15
    # via -r requirements.in
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.35.0
    # via altair
numpy==2.2.5
    # via
    #   -r requirements.in
    #   langchain-community
    #   pandas
    #   pydeck
    #   streamlit
openai==1.75.0
    # via
    #   langchain-openai
    #   litellm
orjson==3.10.16
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   huggingface-hub
    #   langchain-core
    #   langfuse
    #   langsmith
    #   marshmallow
    #   plotly
    #   pytest
    #   streamlit
pandas==2.2.3
    # via
    #   -r requirements.in
    #   pymilvus
    #   streamlit
pillow==11.2.1
    # via streamlit
plotly==5.19.0
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.4
    # via
    #   pymilvus
    #   streamlit
pyarrow==19.0.1
    # via streamlit
pycparser==2.22
    # via cffi
pycryptodome==3.22.0
    # via minio
pydantic==2.11.3
    # via
    #   langchain
    #   langchain-core
    #   langfuse
    #   langsmith
    #   litellm
    #   openai
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pymilvus==2.5.6
    # via -r requirements.in
pymupdf==1.25.5
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   litellm
    #   pydantic-settings
    #   pymilvus
python-magic==0.4.27
    # via -r requirements.in
python-magic-bin==0.4.14
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langfuse
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.40
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
streamlit==1.44.1
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   -r requirements.in
    #   langchain-community
    #   langchain-core
    #   plotly
    #   streamlit
tiktoken==0.9.0
    # via
    #   langchain-openai
    #   litellm
tokenizers==0.21.1
    # via litellm
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
typing-extensions==4.13.2
    # via
    #   altair
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   minio
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
ujson==5.10.0
    # via pymilvus
urllib3==2.4.0
    # via
    #   minio
    #   requests
watchdog==6.0.0
    # via streamlit
wrapt==1.17.2
    # via langfuse
yarl==1.20.0
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
zstandard==0.23.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# setuptools
